str_to_ipaddr() had a bug; use inet_pton() instead; See: https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6844#c27
Also: Improve handling of an invalid value for the 'proxy IP Address'
preference. If the address is invalid, setting the preference
to "" in proto_reg_handoff...() doesn't work well.
Instead, issue a g_warning for an invalid IP address when it'
s processed in proto_reg_handoff...()
(AFAIK, there's not a way for a dissector to validate the semantic
value of a preference and report an error).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@43180 f5534014-38df-0310-8fa8-9805f1628bb7