Move a bunch of common code in the file chooser dialog loops into a routine to run a file chooser dialog.
On Windows, add to that code to resolve shell links ("shortcuts"),
adopted from the Sylpheed mail reader.  (That code requires a pile of
OLE stuff, so link with the OLE library.)

Make it a bit easier to configure ui/gtk/capture_file_dlg.c to use GTK+
dialogs on Windows, but continue to default to using the Win32 dialogs,
at least for now.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@43439 f5534014-38df-0310-8fa8-9805f1628bb7