Don't ask for information about an interface greater than the maximum interface for which we have information.
Fixes bug 7467.

Should also cause an error message, rather than an unreadable capture
file, to be produced for the cases in bug 7381.  (This isn't a fix for
bug 7381; it's arguably an improvement, in the sense that a circuit
breaker tripping or a fuse blowing for a short circuit is an improvement
over a fire starting, but it's not a *fix*, any more than a circuit
break or fuse *fixes* the short circuit.)

#BACKPORT


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@43657 f5534014-38df-0310-8fa8-9805f1628bb7