Add a routine that, given a set of packet encapsulation types, returns the per-file encapsulation type needed to write out a set of packets with all those encapsulation types.  If there's only one such encapsulation type, that's the type, otherwise WTAP_ENCAP_PER_PACKET is needed.  Use that in wtap_dump_can_write_encaps().
Also use it in cf_save_packets() and cf_export_specified_packets(), so
that we can write out files with WTAP_ENCAP_PER_PACKET as the file
encapsulation type and only one actual per-packet encapsulation type in
some cases where that failed before.  This fixes the case that showed up
in bug 7505, although there are other cases where we *could* write out a
capture in a given file format but won't be able to do so; fixing those
will take more work.

#BACKPORT

(Note: this adds a routine to libwiretap, so, when backported, the
*minor* version of the library should be increased.  Code that worked
with the version of the library prior to this change will continue to
work, so there's no need to change the *major* version of the library.)


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@43847 f5534014-38df-0310-8fa8-9805f1628bb7