Fix GCC build.
I'm not exactly sure what's up with the v150fw dissector, since
it defines a heuristic which is never used and registers in no tables.
The comments indicate it's carried over rtp, but rtp doesn't have a
heuristic table. Revision 43887 caused rtp to load a handle to v150fw,
but I've done a quick scan and I don't think it ever actually does anything
with the handle.

For now, #if 0 out the heuristic function to avoid compile errors. A better
fix is needed though.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@43890 f5534014-38df-0310-8fa8-9805f1628bb7