From Stephen Donnelly via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7500 :
The default colorfilter for "Bad Checksum" does not include the Ethernet Frame
Check Sequence (FCS) test.

It seems reasonable that it should be included in this filter rule. It is only
indicated if the Ethernet dissector "Validate the Ethernet checksum if
possible" preference is set and the FCS is believed to be present (via wiretap
heuristic/dissector preference).


From me: re-order the list of bad checksum fields.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@44010 f5534014-38df-0310-8fa8-9805f1628bb7