Correct a number in a value_string.
Spotted in a profile that match_strval_ext() was calling
match_strval_linear() a lot (this value_string has 1476 entries - it was
epensive!).

Found error by adding temp debug to _match_strval_ext_init() - do we
ever expect the entries not to be in order?  Am wondering if the warning
should be committed.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@44576 f5534014-38df-0310-8fa8-9805f1628bb7