*NIX (autofoo and cmake) part of the fix for https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6134 :
Make VERSION_MAJOR, VERSION_MINOR, and VERSION_MICRO available in config.h so
(e.g.) dissector writers can #if around them as necessary to make a piece of
code compile with different versions of Wireshark.  (Hopefully VERSION_MICRO
is not important in this respect!)

Windows also defines VERSION_EXTRA which can be used to help identify custom
builds; this is not done for these build methods (yet?).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@44986 f5534014-38df-0310-8fa8-9805f1628bb7