Patch from Florian Fainelli Fix bug #7772: dissect only the number of active HomePlug AV subcarriers
The dissector currently dissects all 1156 sub-carriers available in the Tone
Map Characteristisc Confirmation frame no matter what is the number of active
ones. 

This is not valid, because the frame only contains the number of active
sub-carriers, even though it is padded to 1156 sub-carriers. This is also an
issue for newest HomePlug AV devices with a 500Mbits/sec PHY rate, because
those will have up to 2690 sub-carriers, and we would only dissect the first
1156 active ones.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@45381 f5534014-38df-0310-8fa8-9805f1628bb7