Based on patch from Bill Parker, check for ws_fopen() failing.
The rest of the code already checks for namelist being null, so it's OK
to call build_tree_view() - and it may be necessary to do so, as the
code below might assume that's worked.

However, we should not call fclose() on namelist if it's null; the C89
standard, at least, does *not* assert that fclose(NULL) is a no-op.

Clean up indentation a little bit.

Last part of fix for bug 7824.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@45455 f5534014-38df-0310-8fa8-9805f1628bb7