In GTK3 the settings.ini file serves the same purpose as the gtkrc file. It should be copied to the wireshark executable dir ..\etc\gtk-3.0
I'm not sure if this is the right place to put it.

makefile will be updated later if no one beats me to it or has a better idea where to put the file. I'm not sure if we should strive to make the GTK3 version as equal to the GTK2 version as possible or accept that it looks more like the *nix version.
If we are to use it it needs to be in the nsi installer as well.

git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@45474 f5534014-38df-0310-8fa8-9805f1628bb7