If we see an SHB after we've read the first SHB, report that as an unsupported feature.
If we see an IDB after all the IDBs at the beginning of the file,
process it.  Fixes bug 7851.

Get rid of unused read_idbs flag in pcapng_t structure.  (Also, as per
the above, just because we've read all the IDBs at the beginning of the
section, that doesn't necessarily mean we've read all the IDBs in the
section.)

Fix some places where we reject SPBs.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@45495 f5534014-38df-0310-8fa8-9805f1628bb7