Pull the length-and-caplen setting operations in the read routines into the macros.
Note why we don't pull the time setting operations into the macros (it's
because that would mean the macros can't be used in the seek-read
routines, as calculating time stamps requires accumulating the time
stamps in a sequential pass through the packets, given that the records
have time stamps relative to the previous packet).

Fix what is probably a typo in the definitions of some macros (which
happen to work without the fix because "fp" and "pp" are both local
variable names as well as macro arguments).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@45705 f5534014-38df-0310-8fa8-9805f1628bb7